Refactoring
- Move Details out of "tables" directory
